school-event-management-system2/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-js-asset==2.1.0
django-model-utils==4.3.1
gunicorn==21.2.0
kombu==5.3.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
